+++ lib/src/eval/compiler/helpers/return.dart
@@ -20,11 +20,25 @@ StatementInfo doReturn(CompilerContext ctx, AlwaysReturnType expectedReturnType,
       final ta = expectedReturnType.type?.specifiedTypeArgs;
       final expected = (ta?.isEmpty ?? true) ? EvalTypes.dynamicType : ta![0];
       var _value = value.boxIfNeeded(ctx);
+
       if (!_value.type.isAssignableTo(ctx, expected)) {
+        if (_value.type.isAssignableTo(ctx, TypeRef.stdlib(ctx, 'dart:core', 'Future'))) {
+          final vta = _value.type.specifiedTypeArgs;
+          final vtype = vta.isEmpty ? EvalTypes.dynamicType : vta[0];
+          if (vtype.isAssignableTo(ctx, expected)) {
+            final _completer = ctx.lookupLocal('#completer')!;
+            final awaitOp = Await.make(_completer.scopeFrameOffset, _value.scopeFrameOffset);
+            ctx.pushOp(awaitOp, Await.LEN);
+            ctx.pushOp(PushReturnValue.make(), PushReturnValue.LEN);
+            final result = Variable.alloc(ctx, EvalTypes.dynamicType);
+            ctx.pushOp(ReturnAsync.make(result.scopeFrameOffset, _completer.scopeFrameOffset), ReturnAsync.LEN);
+            return StatementInfo(-1, willAlwaysReturn: true);
+          }
+        }
         throw CompileError('Cannot return ${_value.type} (expected: $expected)');
       }
       final _completer = ctx.lookupLocal('#completer')!;
-      ctx.pushOp(ReturnAsync.make(_value.scopeFrameOffset, _completer.scopeFrameOffset), Return.LEN);
+      ctx.pushOp(ReturnAsync.make(_value.scopeFrameOffset, _completer.scopeFrameOffset), ReturnAsync.LEN);
       return StatementInfo(-1, willAlwaysReturn: true);
     }